requirements.txt
aiofiles==23.2.1
aiomysql==0.2.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-sqlalchemy==1.0.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
coverage==7.4.4
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.4.0
fastapi==0.110.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyMySQL==1.1.0
pypng==0.20220715.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
qrcode==7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
uvicorn==0.29.0
validators==0.24.0
markdown2
pyjwt
minio==7.2.0
Pillow==10.2.0